refactor: 폼 요소 생성 메서드들을 유틸 함수로 분리
